Can process samples selectively; modulate heatmap parameters (v0.2.2-kneaddata+6)
The user can use the --only_samples option to specify a file
listing the samples to process. This is useful for when additional
samples have been added to the project space, and Nextflow -resume
wouldn't work. (Nextflow -resume will not work if the pipeline code
has changed.)

Heatmap parameters are modulated to better accomodate small sample
sizes.